+++ ReleaseNotes.md
@@ -1,4 +1,3 @@
 # Bluedot ReactNative Minimal Integration
 
-- Update UI to enable/disable background location updates on IOS when the app is in While in use location permissions.
-- Using [Bluedot plugin 2.2.1](https://github.com/Bluedot-Innovation/Bluedot-React-Native-Plugin) (support react-native@0.69.4)
+- Using [Bluedot plugin 2.3.0](https://github.com/Bluedot-Innovation/Bluedot-React-Native-Plugin) (support react-native@>=0.69.0).