+++ ReleaseNotes.md
@@ -1,3 +1,4 @@
 # Bluedot ReactNative Minimal Integration
 
 - Using [Bluedot plugin 2.3.0](https://github.com/Bluedot-Innovation/Bluedot-React-Native-Plugin) (support react-native@>=0.69.0).
+- Updated notification icon for Android