+++ ReleaseNotes.md
@@ -1,4 +1,5 @@
 # Bluedot ReactNative Minimal Integration
 
-- Using [Bluedot plugin 2.3.0](https://github.com/Bluedot-Innovation/Bluedot-React-Native-Plugin) (support react-native@>=0.69.0).
-- Updated notification icon for Android
+- Using Bluedot plugin 2.3.1
+- Added Zone customData in notifications
+- Fixed notifications